requirements.txt
opencv_python==4.6.0.66
Pillow==9.1.1
pyzbar==0.1.9
